*8537* More obvious Dataverse deposit consent, *8553* Allow external citation
Add data publishing options to suppfile form. Offer submitters the option of treating suppfiles as supplementary (not published) or as data files, published in Dataverse. If data to be published with the submission is already in a repository, submitters can provide an external citation.

Add data citation, where present, to revised suppfile forms.

Add form validator to suppfiles to warn when Dataverse deposit selected but no suppfile has been uploaded.

Add form validator to suppfile form to warn when author has provided an external citation *and* has chosen to deposit in Dataverse.

Display external citation, if present & submission does not have an associated Dataverse study, on article landing page & in submission summary table.